scripts/requirements.txt
load_dotenv
jinja2